Tests/LibWeb/Text/input/unclosed-html-element.html

<link>
<html
<head>

<script src="include.js"></script>
<script>
    test(() => {
        println(`HTML element has '<head' attribute: ${document.documentElement.hasAttribute('<head')}`);
    });
</script>
