Tests/LibWeb/Text/expected/unclosed-html-element.txt

HTML element has '<head' attribute: true
